expose new setting
