--- persistent/Database/Persist/Quasi/ParserSettings.hs
@@ -1,8 +1,0 @@
-module Database.Persist.Quasi.ParserSettings
-  ( ParserSettings (..)
-  , ParserWarning
-  , ParserErrorLevel (..)
-  , parserWarningMessage
-  ) where
-
-import Database.Persist.Quasi.ParserSettings.Internal
