+++ persistent/persistent.cabal
@@ -80,10 +80,9 @@ library
     Database.Persist.Names
     Database.Persist.PersistValue
     Database.Persist.Quasi
+    Database.Persist.Quasi.PersistSettings
     Database.Persist.Quasi.Internal
     Database.Persist.Quasi.Internal.ModelParser
-    Database.Persist.Quasi.ParserSettings
-    Database.Persist.Quasi.ParserSettings.Internal
     Database.Persist.Sql
     Database.Persist.Sql.Migration
     Database.Persist.Sql.Types.Internal
